requirements.txt
liac-arff==2.4.0
matplotlib==3.1.3
numpy==1.18.1
pandas==1.1.2
parse==1.15.0
python-Levenshtein==0.12.0
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.1
tqdm==4.46.0
